Champion Attack UI Completed
Champion Attack UI is completed.
You will now see an icon for each of your champion's attack, along with stats of the attack like mana cost, cool down duration, name, keybind. It will indicate that the attack is on cool down and for how long and indicate if you have enough mana to use the attack.